Convert current test suite to `tasty` from `hspec`
