+++ cardano-mempool/tests/Common.hs
@@ -2,6 +2,6 @@ module Common
   ( module X
   ) where
 
-import Test.Hspec as X
-import Test.Hspec.QuickCheck as X
-import Test.QuickCheck as X
+import Test.Tasty as X
+import Test.Tasty.HUnit as X
+import Test.Tasty.QuickCheck as X
