+++ cardano-mempool/tests/Main.hs
@@ -1,11 +1,11 @@
 module Main where
 
-import Spec
 import System.IO (BufferMode (LineBuffering), hSetBuffering, hSetEncoding, stdout, utf8)
-import Test.Hspec
+import Test.Cardano.Memory.PoolTests (poolTests)
+import Test.Tasty
 
 main :: IO ()
 main = do
   hSetBuffering stdout LineBuffering
   hSetEncoding stdout utf8
-  hspec spec
+  defaultMain poolTests
